+++ NEWS.md
@@ -1,3 +1,7 @@
+# BLN 0.9.3 2025-04-10
+## Deprecated
+* function argument A_SOM_LOI in `bln_bbwp_ngw` is deprecated as it was not used by the function
+
 # BLN 0.9.2 2025-04-09
 
 ## Added